+++ weave/trace/op_type.py
@@ -13,7 +13,9 @@ import typing
 from _ast import AsyncFunctionDef, ExceptHandler
 from typing import Any, Callable, Optional, Union, get_args, get_origin
 
-from weave.legacy.weave import artifact_fs, context_state, errors, storage
+from weave import context_state
+from weave.legacy.weave import artifact_fs, errors, storage
+from weave.legacy.weave import context_state as legacy_context_state
 from weave.trace.ipython import (
     ClassNotFoundError,
     get_class_source,
@@ -539,7 +541,7 @@ def load_instance(
     )
 
     sys.path.insert(0, os.path.abspath(module_dir))
-    with context_state.no_op_register():
+    with legacy_context_state.no_op_register():
         try:
             mod = __import__(import_name, fromlist=[module_dir])
         except Exception as e:
